move `tree_route` to client common
